warning eliminated
git-svn-id: https://dune.mathematik.uni-freiburg.de/svn/dune-fem/trunk@2758 4a6c9bf5-c326-0410-8cc6-c7cabc8ddcf0

+++ fem/function/vectorfunction/vectorfunction.hh
@@ -134,7 +134,7 @@ namespace Dune
       dofVector_( &dofVector ),
       freeDofVector_( false )
     {
-      assert( dofVector_->size() == dfSpace.size() );
+      assert( dofVector_->size() == (unsigned int)dfSpace.size() );
     }
 
     inline VectorDiscreteFunction ( const ThisType &other )
